Refactor order analytics: reorganize imports, update query key, and define OrderAnalyticsResponse in api route
